requirements.txt
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
google-generativeai
pypdf
docx2txt
fastapi
uvicorn
